requirements.txt
matplotlib==3.9.0
openai==1.30.5
python_louvain==0.16
mobsfscan==0.3.8
networkx==3.3
javalang==0.13.0
jsonpickle==3.0.4

tree-sitter == 0.21.3
pandas >= 1.3.3
GitPython==3.1.43
deepdiff >=6.2.1
setuptools==65.5.0
tqdm==4.66.4